Refactor shader handling code into a class
This class also automatically frees the shader when in the destructor, the class is also non-copyable to avoid calling glDeleteProgram if there were to be multiple copies of objects of this class, this class also abstracts uniform setting in a nice way that doesn't compromise performance (no forcing of calling glGetUniformLocation every time we set the uniform we can call Shader::get_uniform_setter once to get a callable that sets the uniform, really nice pattern),

Shader::get_uniform_setter is also a templated method, because otherwise we would not have been able to overload it for different return types (the method returns a callable that accepts a vec3 or mat4 for example), so we use template specialization instead.

It would have better if we were able to mix const ref and value type specialization (maybe there is a way that I do not know), but at the moment the template method returns a callable that accepts a const ref of value type,
it is a pity for small data type like float, but I highly believe the compiler will optimize such things in Release mode and so.